分かりやすくした。
