+++ src/VectorNetworkProject/TheMix/event/game/TheEndGameEvent.php
@@ -1,0 +1,31 @@
+<?php
+/**
+ * Copyright (c) 2018 VectorNetworkProject. All rights reserved. MIT license.
+ *
+ * GitHub: https://github.com/VectorNetworkProject/TheMix
+ * Website: https://www.vector-network.tk
+ */
+
+namespace VectorNetworkProject\TheMix\event;
+
+
+use pocketmine\event\Listener;
+use pocketmine\Server;
+use pocketmine\utils\TextFormat;
+use VectorNetworkProject\TheMix\game\DefaultConfig;
+use VectorNetworkProject\TheMix\game\event\game\GameWinEvent;
+use VectorNetworkProject\TheMix\task\ResetGameTask;
+use VectorNetworkProject\TheMix\TheMix;
+
+class TheEndGameEvent implements Listener
+{
+    public function event(GameWinEvent $event)
+    {
+        if (DefaultConfig::isDev()) {
+            $event->setCancelled();
+            return;
+        }
+        TheMix::getInstance()->getScheduler()->scheduleDelayedTask(new ResetGameTask(), 30 * 20);
+        Server::getInstance()->broadcastMessage('===<');
+    }
+}
